requirements.txt
Flask
python-dotenv
Pillow
google-generativeai
